HTTP: Wrap HTTP Get requests in a Perl timeout
Although we are passing a timeout value to HTTP::Tiny there exist some
situations where the delay is happening in a part of the HTTPS
transaction that is not caught by HTTP::Tiny.

This change wraps the entire get request in a alarm timeout set to 1/10
of a second longer than the HTTP::Tiny timeout, this is shown to catch
the problem requests.

Also, maintain an internal (to the instance) list of timed out servers
so we don't for example, send a request for an indicator when the
request for the VMC to the same server already timed out.

Additionally, fix a warning on mark type display where there is no mark
type (for example, when the fetch timed out).